requirements.txt
annotated-types==0.5.0
anyio==3.7.1
click==8.1.7
exceptiongroup==1.1.3
fastapi==0.101.1
h11==0.14.0
idna==3.4
mysql-connector-python==8.1.0
protobuf==4.21.12
pycryptodome==3.18.0
pydantic==2.3.0
pydantic-settings==2.0.3
pydantic_core==2.6.3
python-dotenv==1.0.0
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.7.1
uvicorn==0.23.2
